DAH-1122 Add component for application date, add utils to check if a listing is open for applications and whether a listing is a rental
